abdullah R is evil

+++ restaurants/views.py
@@ -59,6 +59,8 @@ def restaurant_detail(request, restaurant_id):
     return render(request, 'detail.html', context)
 
 def restaurant_create(request):
+    if request.user.is_anonymous:
+        return redirect('signup')
     form = RestaurantForm()
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES)
@@ -75,6 +77,8 @@ def restaurant_create(request):
 def item_create(request, restaurant_id):
     form = ItemForm()
     restaurant = Restaurant.objects.get(id=restaurant_id)
+    if not request.user.is_staff and restaurant.owner != request.user:
+        return redirect('errorpage')
     if request.method == "POST":
         form = ItemForm(request.POST)
         if form.is_valid():
@@ -91,6 +95,8 @@ def item_create(request, restaurant_id):
 def restaurant_update(request, restaurant_id):
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
     form = RestaurantForm(instance=restaurant_obj)
+    if not request.user.is_staff and restaurant_obj.owner != request.user:
+        return redirect('errorpage')
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
         if form.is_valid():
@@ -103,6 +109,14 @@ def restaurant_update(request, restaurant_id):
     return render(request, 'update.html', context)
 
 def restaurant_delete(request, restaurant_id):
+    if not request.user.is_staff:
+        return redirect('errorpage')
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
     restaurant_obj.delete()
     return redirect('restaurant-list')
+
+def NoAccessPage(request):
+    context ={
+        'msg': 'You Have No Access'
+    }
+    return render(request, "redirect.html", context)
